deepfake_service/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
face_recognition
Pillow
